requirements.txt
numpy==2.1.3
openslide-python==1.4.1
pillow==11.0.0
shapely==2.0.6
